requirements.txt
Pillow>=9.0.0
matplotlib>=3.0.0
numpy>=1.20.0
scikit-learn>=1.0.0
